Move Footer to components
It didn't make sense to put it there anyway
